+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New command `report tag` to list tags with time statistics (#93)
 * Add flag `--dry` to `delete` command for dry-run (#96)
+* New command `move project` to move projects between workspaces (#99)
 
 ### Bugfixes